# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: BJR API V1
  description: API specification for the BJR job server.
  version: v1
components:
  schemas:
...
    TodaysStats:
      type: object
      properties:
        total_jobs:
          type: integer
          description: The total number of jobs that are scheduled.
        total_jobs_trend:
          type: number
          format: float
          description: The trend direction of total_jobs.
        enabled_jobs:
          type: integer
          description: The total number of jobs that are enabled. This will be a subset
            of total_jobs.
        enabled_jobs_trend:
          type: number
          format: float
          description: The trend direction of enabled_jobs.
        run_jobs:
          type: integer
          description: The number of jobs that have been run today.
        run_jobs_trend:
          type: number
          format: float
          description: The trend direction of run_jobs.
        failed_jobs:
          type: integer
          description: The number of jobs that have failed today.
        failed_jobs_trend:
          type: number
          format: float
          description: The trend direction of failed_jobs.
        avg_job_runtime:
          type: number
          format: float
          description: The average job run time.
        avg_job_runtime_trend:
          type: number
          format: float
          description: The trend direction of avg_job_runtime.
        max_job_runtime:
          type: number
          format: float
          description: The max job run time.
        max_job_runtime_trend:
          type: number
          format: float
          description: The trend direction of max_job_runtime.
        min_job_runtime:
          type: number
          format: float
          description: The min job run time.
        min_job_runtime_trend:
          type: number
          format: float
          description: The trend direction of min_job_runtime.
        avg_job_lag:
          type: number
          format: float
          description: The average job lag.
        avg_job_lag_trend:
          type: number
          format: float
          description: The trend direction of avg_job_lag.
        max_job_lag:
          type: number
          format: float
          description: The max job lag.
        max_job_lag_trend:
          type: number
          format: float
          description: The trend direction of max_job_lag.
        min_job_lag:
          type: number
          format: float
          description: The min job lag.
        min_job_lag_trend:
          type: number
          format: float
          description: The trend direction of min_job_lag.
    TodaysStatsMessage:
      type: object
      properties:
        message:
          type: string
          description: The status message returned from the API call.
        is_error:
          type: boolean
          description: True if there was an error performing the API call.
        object_type:
          type: string
          description: The type of object being returned.
        status_code:
          type: integer
          description: The HTTP status code returned.
        object:
          "$ref": "#/components/schemas/TodaysStats"
...
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
      bearerFormat: JWT
paths:
...
  "/job_server_api/todays_stats":
    get:
      summary: Todays Stats
      description: Get the high level job statistics for today
      tags:
      - JobServer
      operationId: todaysStats
      security:
      - bearerAuth: []
      parameters:
      - name: timezone
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Get the high level job statistics for today
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TodaysStatsMessage"
...